Repackaged files, now client package created.
