+++ src/main/java/Client/UI/OthersExistingRequests.java
@@ -1,16 +1,10 @@
-package UI;
+package Client.UI;
 
 import Controller.ControllerInterface;
-import Controller.DataPresenter;
 import Controller.IFetch;
-import Controller.ServiceController;
 
 import javax.swing.*;
-import javax.swing.event.ListSelectionEvent;
-import javax.swing.event.ListSelectionListener;
 import java.awt.*;
-import java.awt.event.ActionEvent;
-import java.awt.event.ActionListener;
 
 public class OthersExistingRequests extends JFrame {
     private JPanel mainPanel;
